Move CAS1 Seed Job/Tests into cas1 package
