+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/integration/cas1/seed/SeedApprovedPremisesOfflineApplicationsTest.kt
@@ -1,12 +1,13 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.integration
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.seed
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SeedFileType
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ServiceName
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.SeedTestBase
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.CsvBuilder
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.OfflineApplicationsSeedCsvRow
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.OfflineApplicationsSeedCsvRow
 
 @TestInstance(TestInstance.Lifecycle.PER_METHOD)
 class SeedApprovedPremisesOfflineApplicationsTest : SeedTestBase() {
