+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/integration/cas1/seed/SeedCasUpdateEventNumberTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.seed
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -21,16 +21,15 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.Premise
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.ProbationArea
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.Region
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.Team
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Gender
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SeedFileType
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.StaffUserDetailsFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.SeedTestBase
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.SeedCasUpdateEventNumberTest.CONSTANTS.NEW_CONVICTION_ID
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.SeedCasUpdateEventNumberTest.CONSTANTS.NEW_EVENT_NUMBER
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.SeedCasUpdateEventNumberTest.CONSTANTS.NEW_OFFENCE_ID
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.SeedCasUpdateEventNumberTest.CONSTANTS.OLD_CONVICTION_ID
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.SeedCasUpdateEventNumberTest.CONSTANTS.OLD_EVENT_NUMBER
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.SeedCasUpdateEventNumberTest.CONSTANTS.OLD_OFFENCE_ID
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.seed.SeedCasUpdateEventNumberTest.CONSTANTS.NEW_CONVICTION_ID
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.seed.SeedCasUpdateEventNumberTest.CONSTANTS.NEW_EVENT_NUMBER
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.seed.SeedCasUpdateEventNumberTest.CONSTANTS.NEW_OFFENCE_ID
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.seed.SeedCasUpdateEventNumberTest.CONSTANTS.OLD_CONVICTION_ID
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.seed.SeedCasUpdateEventNumberTest.CONSTANTS.OLD_EVENT_NUMBER
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.cas1.seed.SeedCasUpdateEventNumberTest.CONSTANTS.OLD_OFFENCE_ID
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.givens.`Given a User`
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.givens.`Given an Offender`
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ApprovedPremisesApplicationEntity
